add filter functionality for datasets and clear all for analytics
